requirements.txt
boto==2.44.0
dj-database-url==0.4.1
Django==1.10.4
django-storages==1.5.1
django-taggit==0.21.3
django-taggit-serializer==0.1.5
djangorestframework==3.5.3
gunicorn==19.6.0
psycopg2==2.6.2
Pygments==2.1.3
whitenoise==3.2.2
